finish test refactor
