+++ hub/test/src/common.js
@@ -1,0 +1,9 @@
+import bitcoin from 'bitcoinjs-lib'
+
+export const testWIFs = [
+  'L4kMoaVivcd1FMPPwRU9XT2PdKHPob3oo6YmgTBHrnBHMmo7GcCf',
+  'L3W7EzxYNdG3kBjtQmhKEq2iiZAwpiKEwMobXdaY9xueSUFPkQeH',
+  'KwzzsbVzMekdj9myzxojsgT6DQ6yRQKbWqSXQgo1YKsJcvFJhtRr',
+  'KxYYiJg9mJpCDHcyYMMvSfY4SWGwMofqavxG2ZyDNcXuY7ShBknK']
+export const testPairs = testWIFs.map(x => bitcoin.ECPair.fromWIF(x))
+export const testAddrs = testPairs.map(x => x.getAddress())
