+++ hub/test/src/index.js
@@ -1,6 +1,7 @@
 import { testAuth } from './testAuth.js'
 import { testServer } from './testServer.js'
 import { testDrivers } from './testDrivers.js'
+import { testHttp } from './testHttp.js'
 
 
 
@@ -8,3 +9,4 @@ testServer()
 testAuth()
 testDrivers()
 
+testHttp()
