+++ hub/test/src/testAuth.js
@@ -6,13 +6,7 @@ import * as auth from '../../lib/server/authentication'
 import * as errors from '../../lib/server/errors'
 
 
-const testWIFs = [
-  'L4kMoaVivcd1FMPPwRU9XT2PdKHPob3oo6YmgTBHrnBHMmo7GcCf',
-  'L3W7EzxYNdG3kBjtQmhKEq2iiZAwpiKEwMobXdaY9xueSUFPkQeH',
-  'KwzzsbVzMekdj9myzxojsgT6DQ6yRQKbWqSXQgo1YKsJcvFJhtRr',
-  'KxYYiJg9mJpCDHcyYMMvSfY4SWGwMofqavxG2ZyDNcXuY7ShBknK']
-const testPairs = testWIFs.map(x => bitcoin.ECPair.fromWIF(x))
-const testAddrs = testPairs.map(x => x.getAddress())
+import { testPairs, testAddrs} from './common'
 
 export function testAuth() {
   test('authentication legacy/regression', (t) => {
