+++ hub/test/src/testDrivers.js
@@ -3,8 +3,6 @@ import test  from 'tape'
 import proxyquire from 'proxyquire'
 import FetchMock from 'fetch-mock'
 
-import request from 'supertest'
-import bitcoin from 'bitcoinjs-lib'
 import fs from 'fs'
 
 import { Readable, Writable } from 'stream'
@@ -12,22 +10,14 @@ import { Readable, Writable } from 'stream'
 import * as auth from '../../lib/server/authentication'
 import * as errors from '../../lib/server/errors'
 
-import { makeHttpServer } from '../../lib/server/http.js'
+import { testPairs, testAddrs} from './common'
 
-let azConfigPath = process.env.AZ_CONFIG_PATH
-let awsConfigPath = process.env.AWS_CONFIG_PATH
-let diskConfigPath = process.env.DISK_CONFIG_PATH
-let gcConfigPath = process.env.GC_CONFIG_PATH
+const azConfigPath = process.env.AZ_CONFIG_PATH
+const awsConfigPath = process.env.AWS_CONFIG_PATH
+const diskConfigPath = process.env.DISK_CONFIG_PATH
+const gcConfigPath = process.env.GC_CONFIG_PATH
 
-const testWIFs = [
-  'L4kMoaVivcd1FMPPwRU9XT2PdKHPob3oo6YmgTBHrnBHMmo7GcCf',
-  'L3W7EzxYNdG3kBjtQmhKEq2iiZAwpiKEwMobXdaY9xueSUFPkQeH',
-  'KwzzsbVzMekdj9myzxojsgT6DQ6yRQKbWqSXQgo1YKsJcvFJhtRr',
-  'KxYYiJg9mJpCDHcyYMMvSfY4SWGwMofqavxG2ZyDNcXuY7ShBknK']
-const testPairs = testWIFs.map(x => bitcoin.ECPair.fromWIF(x))
-const testAddrs = testPairs.map(x => x.getAddress())
-
-function addMockFetches(prefix, dataMap) {
+export function addMockFetches(prefix, dataMap) {
   dataMap.forEach( item => {
     FetchMock.get(`${prefix}${item.key}`, item.data)
   })
@@ -42,7 +32,7 @@ function readStream(input, contentLength, callback) {
   })
 }
 
-function makeMockedAzureDriver() {
+export function makeMockedAzureDriver() {
   const dataMap = []
   let bucketName = ''
   const createContainerIfNotExists = function(newBucketName, options, cb) {
@@ -395,94 +385,9 @@ function testGcDriver() {
   })
 }
 
-function testHttpPost() {
-  let config = {
-    "azCredentials": {
-      "accountName": "mock-azure",
-      "accountKey": "mock-azure-key"
-    },
-    "bucket": "spokes"
-  }
-  let mockTest = true
-
-  if (azConfigPath) {
-    config = JSON.parse(fs.readFileSync(azConfigPath))
-    config.driver = 'azure'
-    mockTest = false
-  }
-
-  let dataMap = []
-  let AzDriver
-  const azDriverImport = '../../lib/server/drivers/AzDriver'
-  if (mockTest) {
-    const mockedObj = makeMockedAzureDriver()
-    dataMap = mockedObj.dataMap
-    AzDriver = mockedObj.AzDriver
-    config.driverClass = AzDriver
-  } else {
-    AzDriver = require(azDriverImport)
-  }
-
-  test('handle request', (t) => {
-    let app = makeHttpServer(config)
-    let sk = testPairs[1]
-    let fileContents = sk.toWIF()
-    let blob = Buffer(fileContents)
-
-    let address = sk.getAddress()
-    let path = `/store/${address}/helloWorld`
-    let listPath = `/list-files/${address}`
-    let prefix = ''
-    let authorizationHeader = ''
-
-    request(app).get('/hub_info/')
-      .expect(200)
-      .then((response) => {
-        prefix = JSON.parse(response.text).read_url_prefix
-        const challenge = JSON.parse(response.text).challenge_text
-        const authPart = auth.V1Authentication.makeAuthPart(sk, challenge)
-        return `bearer ${authPart}`
-      })
-      .then((authorization) => {
-        authorizationHeader = authorization
-        return request(app).post(path)
-            .set('Content-Type', 'application/octet-stream')
-            .set('Authorization', authorization)
-            .send(blob)
-            .expect(202)
-      })
-      .then((response) => {
-        if (mockTest) {
-          addMockFetches(prefix, dataMap)
-        }
-
-        let url = JSON.parse(response.text).publicURL
-        t.ok(url, 'Must return URL')
-        console.log(url)
-        fetch(url)
-          .then(resp => resp.text())
-          .then(text => t.equal(text, fileContents, 'Contents returned must be correct'))
-      })
-      .catch((err) => t.false(true, `Unexpected err: ${err}`))
-      .then(() => request(app).post(listPath)
-            .set('Content-Type', 'application/json')
-            .set('Authorization', authorizationHeader)
-            .expect(202)
-      )
-      .then((filesResponse) => {
-        const files = JSON.parse(filesResponse.text)
-        t.equal(files.entries.length, 1, 'Should return one file')
-        t.equal(files.entries[0], 'helloWorld', 'Should be helloworld')
-        t.ok(files.hasOwnProperty('page'), 'Response is missing a page')
-      })
-      .then(() => { FetchMock.restore(); t.end() })
-  })
-}
-
 export function testDrivers() {
   testAzDriver()
   testS3Driver()
   testDiskDriver()
   testGcDriver()
-  testHttpPost()
 }
